+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.5
+
+### Enhancements
+* Add `allow_members_to_invite` to workspace settings
+
+### Bug Fixes
+* Fix issues with API token
+* Fix issues with updating workspace settings
+
 ## 0.2.3
 
 ### Enhancements